Enforce quality guardrails
