+++ .github/workflows/test.yml
@@ -36,6 +36,9 @@ jobs:
       - name: Run checkJs pilot
         run: npm run typecheck:checkjs
 
+      - name: Run quality guardrails
+        run: npm run quality
+
       - name: Cache Playwright browsers
         uses: actions/cache@v5
         with:
