+++ service-worker.js
@@ -78,6 +78,7 @@ const APP_SHELL = [
   '/js/app-event-listeners.js',
   '/js/startup-orchestrator.js',
   '/js/schema.js',
+  '/js/schema-environment.js',
   '/js/constants.js',
   '/js/state.js',
   '/js/utils.js',
@@ -115,6 +116,8 @@ const APP_SHELL = [
   '/js/context-source-registry.js',
   '/js/context-card-summaries.js',
   '/js/context-card-editor-ui.js',
+  '/js/context-card-dashboard-ai.js',
+  '/js/context-card-dashboard-ai-actions.js',
   '/js/context-card-medical-history-editor.js',
   '/js/context-card-lifestyle-editors.js',
   '/js/focus-card.js',
@@ -250,6 +253,7 @@ const APP_SHELL = [
   '/js/dna-actions.js',
   '/js/dna-genotype.js',
   '/js/dna.js',
+  '/js/dna-window-bindings.js',
   '/js/dna-mtdna.js',
   '/js/hardware.js',
   '/js/sync.js',
@@ -378,6 +382,7 @@ const APP_SHELL = [
   '/js/light-device-ai-analysis.js',
   '/js/light-sun-ai-hooks.js',
   '/js/light-devices.js',
+  '/js/light-devices-actions.js',
   '/js/light-devices-store.js',
   '/js/light-device-session-engine.js',
   '/js/light-device-setup-modal.js',
