+++ tsconfig.checkjs.json
@@ -70,6 +70,7 @@
     "js/chat.js",
     "js/client-list.js",
     "js/constants.js",
+    "js/context-card-dashboard-ai-actions.js",
     "js/context-card-dashboard-ai.js",
     "js/context-card-editor-ui.js",
     "js/context-card-health-dots.js",
@@ -88,6 +89,7 @@
     "js/data-merge.js",
     "js/data.js",
     "js/dna.js",
+    "js/dna-window-bindings.js",
     "js/emf.js",
     "js/export-report-builder.js",
     "js/export-report-html.js",
@@ -122,6 +124,7 @@
     "js/light-device-session-engine.js",
     "js/light-device-session-modal.js",
     "js/light-device-setup-modal.js",
+    "js/light-devices-actions.js",
     "js/light-devices.js",
     "js/light-devices-store.js",
     "js/light-env.js",
@@ -351,6 +354,7 @@
     "js/recommendation-actions.js",
     "js/recommendations-region.js",
     "js/schema.js",
+    "js/schema-environment.js",
     "js/settings-data.js",
     "js/settings-provider-bridge.js",
     "js/settings-sync-panel.js",
