+++ tsconfig.json
@@ -70,6 +70,7 @@
     "js/commit-hash.js",
     "js/compare-correlations.js",
     "js/constants.js",
+    "js/context-card-dashboard-ai-actions.js",
     "js/context-card-dashboard-ai.js",
     "js/context-card-editor-ui.js",
     "js/context-card-health-dots.js",
@@ -88,6 +89,7 @@
     "js/data-merge.js",
     "js/data.js",
     "js/dna.js",
+    "js/dna-window-bindings.js",
     "js/emf.js",
     "js/emf-facade.js",
     "js/export.js",
@@ -119,6 +121,7 @@
     "js/light-device-session-engine.js",
     "js/light-device-session-modal.js",
     "js/light-device-setup-modal.js",
+    "js/light-devices-actions.js",
     "js/light-devices-store.js",
     "js/light-devices.js",
     "js/light-env-actions.js",
@@ -176,6 +179,7 @@
     "js/provider-wallet-delegates.js",
     "js/provider-wallet-panels.js",
     "js/schema.js",
+    "js/schema-environment.js",
     "js/settings.js",
     "js/settings-sync-panel.js",
     "js/settings-agent-access-panel.js",
